物资类型优化1，已被使用物资不允许删除，异步级联更新

+++ src/main/java/com/zxl/materialStorage/controller/materialEnter/MaterialTypeController.java
@@ -1,15 +1,22 @@
 package com.zxl.materialStorage.controller.materialEnter;
 
 import cn.hutool.core.util.ObjectUtil;
+import com.baomidou.mybatisplus.core.conditions.query.QueryWrapper;
 import com.baomidou.mybatisplus.core.toolkit.StringUtils;
 import com.zxl.materialStorage.common.api.ApiResult;
+import com.zxl.materialStorage.model.pojo.MaterialEnter;
 import com.zxl.materialStorage.model.pojo.MaterialType;
+import com.zxl.materialStorage.service.materialEnter.MaterialEnterService;
 import com.zxl.materialStorage.service.materialEnter.MaterialTypeService;
+import com.zxl.materialStorage.util.SystemUtil;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
+import java.util.ArrayList;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 
 /**
  * @className: MaterialTypeController
@@ -23,68 +30,92 @@ import java.util.List;
 public class MaterialTypeController {
     @Autowired
     private MaterialTypeService materialTypeService;
+    @Autowired
+    private MaterialEnterService materialEnterService;
 
     @PostMapping("/insertNewOne")
-    public ApiResult<Object> insertNewOne(@RequestBody MaterialType materialType){
+    public ApiResult<Object> insertNewOne(@RequestBody MaterialType materialType) {
         try {
-            if (ObjectUtil.isNull(materialType)){
+            if (ObjectUtil.isNull(materialType)) {
                 return ApiResult.blank();
             }
             materialTypeService.insertNewOne(materialType);
         } catch (Exception e) {
-            log.error("新增物资类型出错",e);
-            return ApiResult.error(500,"新增物资类型出错",e);
+            log.error("新增物资类型出错", e);
+            return ApiResult.error(500, "新增物资类型出错", e);
         }
         return ApiResult.success();
     }
 
-    //TODO 若存在引用则拒绝删除
     @DeleteMapping("/deleteOne")
-    public ApiResult<Object> deleteOne(@RequestParam(value = "emtId") String emtId){
+    public ApiResult<Object> deleteOne(@RequestParam(value = "emtId") String emtId) {
         try {
-            if (StringUtils.isEmpty(emtId)){
+            if (StringUtils.isEmpty(emtId)) {
                 return ApiResult.blank();
+            }
+            //存在引用，拒绝删除
+            MaterialType byId = materialTypeService.getById(emtId);
+            MaterialEnter existMaterialEnter = materialEnterService.getOne(new QueryWrapper<MaterialEnter>().lambda().eq(MaterialEnter::getEmtNo, byId.getEmtNo()));
+            if (ObjectUtil.isNotNull(existMaterialEnter)) {
+                return ApiResult.error();
             }
             materialTypeService.deleteOne(emtId);
         } catch (Exception e) {
-            log.error("删除物资类型出错",e);
-            return ApiResult.error(500,"删除物资类型出错",e);
+            log.error("删除物资类型出错", e);
+            return ApiResult.error(500, "删除物资类型出错", e);
         }
         return ApiResult.success();
     }
 
-    //TODO 若存在引用则拒绝删除
     @DeleteMapping("/deleteMany")
-    public ApiResult<Object> deleteMany(@RequestBody List<String> emtIdList){
+    public ApiResult<Object> deleteMany(@RequestBody List<String> emtIdList) {
         try {
-            if (ObjectUtil.isEmpty(emtIdList)){
+            if (ObjectUtil.isEmpty(emtIdList)) {
                 return ApiResult.blank();
+            }
+            //存在引用，拒绝删除
+            List<MaterialType> materialTypeList = materialTypeService.listByIds(emtIdList);
+            List<String> emtNoList = new ArrayList<>();
+            for (MaterialType materialType : materialTypeList) {
+                emtNoList.add(materialType.getEmtNo());
+            }
+            Map<String, Object> columnMap = new HashMap<>();
+            columnMap.put("emt_no",emtNoList);
+            List<MaterialEnter> materialEnters = materialEnterService.listByMap(columnMap);
+            if (ObjectUtil.isNotEmpty(materialEnters)){
+                return ApiResult.error();
             }
             materialTypeService.deleteMany(emtIdList);
         } catch (Exception e) {
-            log.error("批量删除物资类型出错",e);
-            return ApiResult.error(500,"批量删除物资类型出错",e);
+            log.error("批量删除物资类型出错", e);
+            return ApiResult.error(500, "批量删除物资类型出错", e);
         }
         return ApiResult.success();
     }
 
-    //TODO 若存在引用，则更新引用
     @PostMapping("/updateOne")
-    public ApiResult<Object> updateOne(@RequestBody MaterialType materialType){
+    public ApiResult<Object> updateOne(@RequestBody MaterialType materialType) {
         try {
-            if (ObjectUtil.isNull(materialType)){
+            if (ObjectUtil.isNull(materialType)) {
                 return ApiResult.blank();
             }
+            //异步更新引用
+            MaterialType byId = materialTypeService.getById(materialType.getEmtId());
+            if (!byId.getEmtNo().equals(materialType.getEmtNo())){
+                materialEnterService.updateEmtNos(byId,materialType);
+            }
+
+            //再更新自己
             materialTypeService.updateOne(materialType);
         } catch (Exception e) {
-            log.error("更新物资类型出错");
-            return ApiResult.error(500,"更新物资类型出错",e);
+            log.error("更新物资类型出错",e);
+            return ApiResult.error(500, "更新物资类型出错", e);
         }
         return ApiResult.success();
     }
 
     @GetMapping("/selectAll")
-    public ApiResult<List<MaterialType>> selectAll(){
+    public ApiResult<List<MaterialType>> selectAll() {
         return ApiResult.success(materialTypeService.selectAll());
     }
 }
